Extract random effects from a gam

Extract what would be the random effects from a mixed model from a gam object. Assumes an mgcv model of the form gam(... + s(g, bs='re')).

extract_ranef(model, re = NULL, tibble = TRUE)

Arguments

model
The mgcv model

re
Which specific coefficients to extract. Currently not implemented.

tibble
Return a tibble or standard default data.frame. Default is TRUE.

Value

A tibble (if tibble = TRUE) or data frame with the random effect, its standard error, and its lower and upper bounds. The bounds are based on a simple normal approximation using the standard error.

Details

Returns a data frame of the the component type, the estimated random effect re, the estimated se, and approximate lower and upper bounds assuming +- 1.96*se. Note that the standard errors are Bayesian estimates (see gamObject, specifically type Vp). The re will only reflect smooth terms whose basis function is 're', i.e. of class random.effect. Others will be ignored.

Note

mgcv strips the level names for 're' smooth terms, so this attempts to get them back. This may not work under every circumstance, but the attempt is made to extract the names of random effect groups based on how they are ordered in the data (which is how the model matrix would be constructed), and in the case of random slopes, detect that second variable in the 're' specification would be the grouping variable. This will not work for continuous x continuous smooths of type 're', but I can't think of a reason why you'd use that given your other options with mgcv.

See also

Examples

library(mgcv)
library(lme4)

lmer_model <- lmer(Reaction ~ Days + (Days || Subject), data = sleepstudy)
ga_model <- gam(Reaction ~ Days + s(Subject, bs = "re") + s(Days, Subject, bs = "re"),
  data = sleepstudy,
  method = "REML"
)

ranef(lmer_model)

#> $Subject
#>     (Intercept)        Days
#> 308   1.5116973   9.3237308
#> 309 -40.3720066  -8.5995358
#> 310 -39.1795117  -5.3880715
#> 330  24.5188072  -4.9686809
#> 331  22.9141930  -3.1939310
#> 332   9.2217762  -0.3084673
#> 333  17.1557243  -0.2871512
#> 334  -7.4516633   1.1159907
#> 335   0.5798353 -10.9062401
#> 337  34.7661741   8.6279628
#> 349 -25.7538155   1.2806250
#> 350 -13.8653871   6.7565202
#> 351   4.9161750  -3.0751926
#> 352  20.9281601   3.5123758
#> 369   3.2584763   0.8730848
#> 370 -26.4756822   4.9838147
#> 371   0.9057286  -1.0053150
#> 372  12.4213189   1.2584805
#> 
#> with conditional variances for “Subject” 

extract_ranef(ga_model)

#> # A tibble: 36 x 6
#>    component group      re    se  lower upper
#>    <fct>     <fct>   <dbl> <dbl>  <dbl> <dbl>
#>  1 Subject   308     1.51   13.3 -24.5   27.5
#>  2 Subject   309   -40.4    13.3 -66.4  -14.3
#>  3 Subject   310   -39.2    13.3 -65.2  -13.2
#>  4 Subject   330    24.5    13.3  -1.51  50.5
#>  5 Subject   331    22.9    13.3  -3.11  48.9
#>  6 Subject   332     9.22   13.3 -16.8   35.2
#>  7 Subject   333    17.2    13.3  -8.87  43.2
#>  8 Subject   334    -7.45   13.3 -33.5   18.6
#>  9 Subject   335     0.579  13.3 -25.4   26.6
#> 10 Subject   337    34.8    13.3   8.74  60.8
#> # … with 26 more rows